PB - MF-287 - Move basic input and rename basic select
