+++ src/patient-registration/input/basic-input/input/basic-input.component.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useField } from 'formik';
-import styles from './../input.css';
+import styles from './../../input.css';
 
 interface BasicInputProps {
   type: string;
