+++ src/patient-registration/input/basic-input/select/basic-select.test.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { render, fireEvent, wait } from '@testing-library/react';
 import { Formik, Form } from 'formik';
-import { SelectInput } from './select-input.component';
+import { BasicSelect } from './basic-select.component';
 
-describe('select input', () => {
-  const setupInput = async () => {
+describe('basic select', () => {
+  const setupSelect = async () => {
     const { getByLabelText } = render(
       <Formik initialValues={{ select: '' }} onSubmit={null}>
         <Form>
-          <SelectInput label="Select" showLabel={true} name="select" options={['A Option', 'B Option']} />
+          <BasicSelect label="Select" showLabel={true} name="select" options={['A Option', 'B Option']} />
         </Form>
       </Formik>,
     );
@@ -16,12 +16,12 @@ describe('select input', () => {
   };
 
   it('exists', async () => {
-    const input = await setupInput();
+    const input = await setupSelect();
     expect(input.type).toEqual('select-one');
   });
 
   it('can input data', async () => {
-    const input = await setupInput();
+    const input = await setupSelect();
     const expected = 'A Option';
 
     fireEvent.change(input, { target: { value: expected } });
